banking_alliance_demo/requirements.txt
snowflake-snowpark-python>=1.14.0
snowflake-connector-python>=3.8.0
pandas>=2.0.0
faker>=24.0.0
python-dateutil>=2.8.0
toml>=0.10.2
pyyaml>=6.0
